requirements.txt
groq>=0.12.0
python-dotenv>=1.0.0
rich>=13.7.0
